requirements.txt
numpy
xarray
netCDF4
pillow
matplotlib
cartopy
tqdm
scipy
geopandas
requests
shapely
pyproj
fiona
